requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
pillow==10.0.1
speechrecognition==3.10.0
pydub==0.25.1
werkzeug==2.3.7